some adjustments
