+++ packages/frontend/codegen.yml
@@ -1,7 +1,7 @@
 overwrite: true
 schema:
   - 'http://localhost:3000/graphql'
-  - 'src/graphql/local-only-schema/schema.gql'
+  - 'src/graphql/local-only/schema.gql'
 documents:
   - 'src/graphql/**/*.gql'
   - 'src/**/*.{ts,tsx,js,jsx,vue}'
@@ -15,4 +15,3 @@ generates:
     config:
       scalars:
         JSONObject: Record<string, unknown>
-        UnknownLocalObject: unknown
