+++ packages/frontend/src/main/lib/core/composables/core.ts
@@ -1,5 +1,8 @@
+import { OverridedMixpanel } from 'mixpanel-browser'
+import { getCurrentInstance, computed } from 'vue'
+import { useQuery } from '@vue/apollo-composable'
+import { IsLoggedInDocument } from '@/graphql/generated/graphql'
 import { ComposableInvokedOutOfScopeError } from '@/main/lib/core/errors/composition'
-import { getCurrentInstance } from 'vue'
 
 /**
  * Get EventHub
@@ -9,4 +12,23 @@ export function useEventHub() {
   if (!vm) throw new ComposableInvokedOutOfScopeError()
 
   return vm.proxy.$eventHub
+}
+
+/**
+ * Get Mixpanel instance (not reactive)
+ */
+export function useMixpanel(): OverridedMixpanel {
+  const vm = getCurrentInstance()
+  if (!vm) throw new ComposableInvokedOutOfScopeError()
+
+  return vm.proxy.$mixpanel
+}
+
+/**
+ * Composable that resolves whether the user is logged in through an Apollo query
+ */
+export function useIsLoggedIn() {
+  const { result } = useQuery(IsLoggedInDocument)
+  const isLoggedIn = computed(() => !!result.value?.user?.id)
+  return { isLoggedIn }
 }
